expose interal types through mlogjs:types
This allows us to use a module declaration for mlogjs:world, without needing to rely on the generated tsconfig to do that
