+++ compiler/lib/traits.d.ts
@@ -1,16 +1,4 @@
-import "./globals";
-
-type LogicSymbols = typeof Items & typeof Liquids & typeof LAccess;
-
-type AsSymbolRecord<T extends Record<string, unknown>> = {
-  -readonly [K in keyof (LogicSymbols | T) as LogicSymbols[K]]: T[K];
-};
-
-export type MutableWithSymbols<T extends Record<string, unknown>> = T &
-  AsSymbolRecord<T>;
-
-export type WithSymbols<T extends Record<string, unknown>> = T &
-  Readonly<AsSymbolRecord<T>>;
+import { WithSymbols } from "mlogjs:types";
 
 declare global {
   interface LiquidHolder
