+++ compiler/lib/util.d.ts
@@ -1,123 +1,137 @@
-import { MutableWithSymbols } from "./traits";
+/// <reference path="./globals.d.ts" />
 
-export type TDrawPrintAlign =
-  | "center"
-  | "top"
-  | "bottom"
-  | "left"
-  | "right"
-  | "topLeft"
-  | "topRight"
-  | "bottomLeft"
-  | "bottomRight";
+declare module "mlogjs:types" {
+  type LogicSymbols = typeof Items & typeof Liquids & typeof LAccess;
 
-export type TRadarFilter =
-  | "any"
-  | "enemy"
-  | "ally"
-  | "player"
-  | "attacker"
-  | "flying"
-  | "boss"
-  | "ground";
+  type AsSymbolRecord<T extends Record<string, unknown>> = {
+    -readonly [K in keyof (LogicSymbols | T) as LogicSymbols[K]]: T[K];
+  };
 
-export type TRadarFilterArray =
-  | [TRadarFilter]
-  | [TRadarFilter, TRadarFilter]
-  | [TRadarFilter, TRadarFilter, TRadarFilter];
+  export type MutableWithSymbols<T extends Record<string, unknown>> = T &
+    AsSymbolRecord<T>;
 
-export type TRadarSort =
-  | "distance"
-  | "health"
-  | "shield"
-  | "armor"
-  | "maxHealth";
+  export type WithSymbols<T extends Record<string, unknown>> = T &
+    Readonly<AsSymbolRecord<T>>;
 
-export type TUnitLocateBuildingGroup =
-  | "core"
-  | "storage"
-  | "generator"
-  | "turret"
-  | "factory"
-  | "repair"
-  | "rally"
-  | "battery"
-  | "reactor";
+  export type TDrawPrintAlign =
+    | "center"
+    | "top"
+    | "bottom"
+    | "left"
+    | "right"
+    | "topLeft"
+    | "topRight"
+    | "bottomLeft"
+    | "bottomRight";
 
-export type TStatusEffect =
-  | "burning"
-  | "freezing"
-  | "unmoving"
-  | "slow"
-  | "wet"
-  | "muddy"
-  | "melting"
-  | "sapped"
-  | "tarred"
-  | "overclock"
-  | "shielded"
-  | "shocked"
-  | "blasted"
-  | "corroded"
-  | "spore-slowed"
-  | "disarmed"
-  | "electrified"
-  | "invincible"
-  | TPermanentStatusEffect;
+  export type TRadarFilter =
+    | "any"
+    | "enemy"
+    | "ally"
+    | "player"
+    | "attacker"
+    | "flying"
+    | "boss"
+    | "ground";
 
-export type TPermanentStatusEffect = "boss" | "overdrive";
+  export type TRadarFilterArray =
+    | [TRadarFilter]
+    | [TRadarFilter, TRadarFilter]
+    | [TRadarFilter, TRadarFilter, TRadarFilter];
 
-type CommonSettableProps = Record<keyof typeof Items, number> & {
-  team: TeamIdentifier;
-  health: number;
-};
+  export type TRadarSort =
+    | "distance"
+    | "health"
+    | "shield"
+    | "armor"
+    | "maxHealth";
 
-interface SettableUnit
-  extends MutableWithSymbols<
-    CommonSettableProps & {
-      x: number;
-      y: number;
-      flag: number;
-      rotation: number;
-      payloadType?: UnitSymbol | BuildingSymbol;
-      speed: number;
-      armor: number;
-    }
-  > {}
+  export type TUnitLocateBuildingGroup =
+    | "core"
+    | "storage"
+    | "generator"
+    | "turret"
+    | "factory"
+    | "repair"
+    | "rally"
+    | "battery"
+    | "reactor";
 
-interface SettableBuilding
-  extends MutableWithSymbols<
-    CommonSettableProps &
-      Record<keyof typeof Liquids, number> & {
-        totalPower: number;
+  export type TStatusEffect =
+    | "burning"
+    | "freezing"
+    | "unmoving"
+    | "slow"
+    | "wet"
+    | "muddy"
+    | "melting"
+    | "sapped"
+    | "tarred"
+    | "overclock"
+    | "shielded"
+    | "shocked"
+    | "blasted"
+    | "corroded"
+    | "spore-slowed"
+    | "disarmed"
+    | "electrified"
+    | "invincible"
+    | TPermanentStatusEffect;
+
+  export type TPermanentStatusEffect = "boss" | "overdrive";
+
+  type CommonSettableProps = Record<keyof typeof Items, number> & {
+    team: TeamIdentifier;
+    health: number;
+  };
+
+  interface SettableUnit
+    extends MutableWithSymbols<
+      CommonSettableProps & {
+        x: number;
+        y: number;
+        flag: number;
+        rotation: number;
+        payloadType?: UnitSymbol | BuildingSymbol;
+        speed: number;
+        armor: number;
       }
-  > {}
+    > {}
 
-export type TLogicLinkNames<T extends string> =
-  `${TLogicLinkName<T>}${TLogicLinkDigit}`;
+  interface SettableBuilding
+    extends MutableWithSymbols<
+      CommonSettableProps &
+        Record<keyof typeof Liquids, number> & {
+          totalPower: number;
+        }
+    > {}
 
-/**
- * Converts a the name of a block into it's logic link variant.
- *
- * The transformation happens as follows:
- *
- * - If the name is all lowercase, return it.
- * - If the name is a single uppercase letter followed by lowercase letters,
- *   return the name in lowercase form.
- * - If the name has "Large" at the end, remove it and try again.
- * - Else, remove one letter from the beggining and try again.
- *
- * Based on [the mindustry source
- * code](https://github.com/Anuken/Mindustry/blob/93daa7a5dcc3fac9e5f40c3375e9f57ae4720ff4/core/src/mindustry/world/blocks/logic/LogicBlock.java#L103-L115).
- */
-export type TLogicLinkName<T extends string> = T extends Lowercase<string>
-  ? T
-  : T extends `${Uppercase<string>}${Lowercase<string>}`
-  ? Lowercase<T>
-  : T extends `${infer Begin}Large`
-  ? TLogicLinkName<Begin>
-  : T extends `${string}${infer Rest}`
-  ? TLogicLinkName<Rest>
-  : never;
+  export type TLogicLinkNames<T extends string> =
+    `${TLogicLinkName<T>}${TLogicLinkDigit}`;
 
-export type TLogicLinkDigit = 1 | 2 | 3 | 4 | 6 | 7 | 8 | 9;
+  /**
+   * Converts a the name of a block into it's logic link variant.
+   *
+   * The transformation happens as follows:
+   *
+   * - If the name is all lowercase, return it.
+   * - If the name is a single uppercase letter followed by lowercase letters,
+   *   return the name in lowercase form.
+   * - If the name has "Large" at the end, remove it and try again.
+   * - Else, remove one letter from the beggining and try again.
+   *
+   * Based on [the mindustry source
+   * code](https://github.com/Anuken/Mindustry/blob/93daa7a5dcc3fac9e5f40c3375e9f57ae4720ff4/core/src/mindustry/world/blocks/logic/LogicBlock.java#L103-L115).
+   */
+  export type TLogicLinkName<T extends string> = T extends Lowercase<string>
+    ? T
+    : T extends `${Uppercase<string>}${Lowercase<string>}`
+    ? Lowercase<T>
+    : T extends `${infer Begin}Large`
+    ? TLogicLinkName<Begin>
+    : T extends `${string}${infer Rest}`
+    ? TLogicLinkName<Rest>
+    : never;
+
+  export type TLogicLinkDigit = 1 | 2 | 3 | 4 | 6 | 7 | 8 | 9;
+}
